requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
gitpython==3.1.40
pygithub==2.1.1
playwright==1.40.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dateutil==2.8.2
